Adding a "More songs" feature to generated playlist.
Implementing the first version of a "More songs" feature. The changes made were mainly adding a parameter "playlistLength" to the generation of a playlist. So far, it hits the Spotify API everytime the user asks for more songs. This isn't necessary as we already have the full list of songs from the first search.